Spinner class in own .js library file
